[api-extractor] Add scenario for #3593, incorrect references to renamed classes
"Incorrect canonical reference to aliased class in .api.json"
https://github.com/microsoft/rushstack/issues/3593

Interface "Options" is renamed to "OptionsRenamed" via
`export { __ as __ }`. This works consistently for *.api.md
and rollup.d.ts as well as the "name" in *.api.json, but not
for the several other references in *.api.json (see git diff
after running api-extractor-scenarios).