[api-extractor] Fix 'export __ as __' part of #3593
"Incorrect canonical reference to aliased class in .api.json"
https://github.com/microsoft/rushstack/issues/3593

Changed the example a bit to be closer to the targeted
package / module layout: "Item" and "Options" are now
default exports of two seperate source files, and
"index" re-exports both, renaming "Options" to
"renamed_Options" (preserving some kind of "namespace").
For this layout, the problem has been fixed.
DeclarationReferenceGenerator now receives a function
to resolve a Symbol to its emit name.
ApiModelGenerator fetches this function from Collector,
which now maintains another Map from Symbol to
CollectorEntity. If for a given Symbol, there is
a CollectorEntity, then its emitName is returned,
the Symbol's name otherwise.
While this change fixes the export { __ as __ }
scenario, for _some_ reason, it changes the output
of one other test, namely ambientNameConflict.
In its *.api.json file, some reference to the
class re-exported as "MyPromise", the name clash
of localFile's "Promise" and the global "Promise"
now leads to "Promise_2" being used instead of
"Promise", the old expected result that was also
wrong, because it should have been "MyPromise",
and "MyPromise" should appear in *.api.json, but
that is something completely different, so I'll
ignore the problem for now and check in the changed
expected *.api.json of ambientNameConflict.